refactor delete account

+++ staff/alba-casas/full-stack/final-project/logic/src/deleteAccount.js
@@ -9,8 +9,11 @@ function deleteAccount(userId, password) {
   validateId(userId);
   validatePassword(password);
 
-  return Job.deleteMany({ userId }).then(() => {
-    return User.deleteOne({ _id: userId, password });
+  return User.findById(userId).then((user) => {
+    if (user.password !== password) throw new Error("Wrong credentials");
+    return Job.deleteMany({ userId }).then(() => {
+      return User.deleteOne({ _id: userId, password });
+    });
   });
 }
